doDeepSleep: changed the datatype of time_us to uint64_t, according to ESP8266 Arduino core 2.4.1 (#656)
As a result it is possible to set durations longer than ~71min.

+++ src/Homie.cpp
@@ -369,7 +369,7 @@ void HomieClass::prepareToSleep() {
   }
 }
 
-void HomieClass::doDeepSleep(uint32_t time_us, RFMode mode) {
+void HomieClass::doDeepSleep(uint64_t time_us, RFMode mode) {
   Interface::get().getLogger() << F("💤 Device is deep sleeping...") << endl;
   Serial.flush();
   ESP.deepSleep(time_us, mode);
